Fixing stuff

+++ src/main_candidate.c
@@ -171,8 +171,8 @@ receive_file(int sockfd, struct sockaddr_in *addr, const char *filename) {
         }
 
         // TODO: do parsing differently maybe, shammt is weird
-        uint16_t opcode = (buffer[0] << 8) | buffer[1];
-        uint16_t received_block_num = (buffer[2] << 8) | buffer[3];
+        uint16_t opcode = (uint16_t)(buffer[0] << 8 | buffer[1]);
+        uint16_t received_block_num = (uint16_t)(buffer[2] << 8 | buffer[3]);
 
         received_block_num = ntohs(received_block_num);
